refactor(findings): extract useOfflineWrite hook
Move the tryEnqueueOffline offline-queue write helper out of useFindings
into a dedicated useOfflineWrite hook. The version-freeze / shouldQueue /
lastKnownVersion semantics are preserved verbatim; useFindings calls the
hook and uses the returned callback at all existing sites. Behavior-
preserving; offline-queue web tests stay green.
